Move optimization constraints params to optimizer package
